requirements.txt
# =============================================================================
# Combined Requirements for WiFi Optimization Simulation Platform
# =============================================================================

# =============================================================================
# SYSTEM DEPENDENCIES (install via package manager BEFORE pip install)
# =============================================================================
#
# Ubuntu/Debian:
#   sudo apt-get install -y librdkafka-dev rapidjson-dev pkg-config
#
# Arch Linux:
#   sudo pacman -S librdkafka rapidjson
#
# macOS:
#   brew install librdkafka rapidjson pkg-config
#
# These are required for:
#   - contrib/kafka-producer (C++ module)
#   - contrib/kafka-consumer (C++ module)
#   - modern-cpp-kafka headers (included in repo, wraps librdkafka)
#
# =============================================================================

# -----------------------------------------------------------------------------
# Core RL Dependencies (for RL/main.py)
# -----------------------------------------------------------------------------
torch>=2.0.0
torch-geometric>=2.3.0
numpy>=1.24.0
networkx>=3.0
scipy>=1.10.0

# -----------------------------------------------------------------------------
# Kafka Integration (for simulation <-> RL communication)
# -----------------------------------------------------------------------------
kafka-python>=2.0.2

# -----------------------------------------------------------------------------
# Visualization Backend (for visualize-data/backend/server.py)
# -----------------------------------------------------------------------------
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# -----------------------------------------------------------------------------
# Data Processing & Utilities
# -----------------------------------------------------------------------------
matplotlib>=3.7.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0

# -----------------------------------------------------------------------------
# Optional: Jupyter Support
# -----------------------------------------------------------------------------
# jupyter>=1.0.0
# ipykernel>=6.25.0
# notebook>=7.0.0
